front-service/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.0
qrcode==7.4.2